+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 Changes to this project are documented in this file. More detail and links can be found in the Declarative Onboarding [Document Revision History](https://clouddocs.f5.com/products/extensions/f5-declarative-onboarding/latest/revision-history.html).
 
+## 1.40.0
+### Added
+- AUTOTOOL-3931: ([GitHub Issue 656](https://github.com/F5Networks/f5-appsvcs-extension/issues/656)): Add WAF Settings configuration support on DO
+- AUTOTOOL-3959: Add routeDomain property to RoutingPrefixList
+- AUTOTOOL-1631: ([GitHub Issue 139](https://github.com/F5Networks/f5-declarative-onboarding/issues/139)): Configure username and password prompts
+
+### Fixed
+- AUTOTOOL-3942: Fix documentation output from newlines in guiSecurityBannerText
+
+### Changed
+
+### Removed
+
 ## 1.39.0
 ### Added